+++ CHANGES.rst
@@ -9,6 +9,7 @@ CHANGE HISTORY
   - Retired Blessed support.
 - Added set_scenes() and draw_next_frame() to allow asynchronous frameworks to
   use Screen.
+- Added support for multi-colour rendering using ${c,a,b} syntax.
 - Added UT framework for testing and CI configurations to run the tests.
 - Fixed bug in restoring console colours on Exit for Windows.
 - Fixed up logic for handling Ctrl keys and documented OS restrictions.